requirements.txt
flask
pillow
requests[socks]
wand
sentry_sdk
redis
numpy
scipy
psycopg2-binary
cairosvg
apng
lottie[GIF]
colour